client/requirements.txt
pvporcupine>=3.0.4
pyaudio>=0.2.14
python-dotenv>=1.0.0
